@gamb re: e90f9a
For fiction, I've recently listened to the first two "Earthsea" books! They're high-fantasy, wizards and dragons and whatnot. I liked them.

For nonfiction, "Caste: The Origins of our Discontents" by Isabel Wilkerson. It argues that racism in the US is best described as a caste system. It's supported with heavy and graphic historical examples though, definitely not happy holiday reading.

I can also recommend Philip Rogaway's "The Moral Character of Cryptography," but it's a paper and not a story. It doesn't require deep cryptography expertise to appreciate (except maybe in part 4, which i haven't yet read)